add button-functionality
