+++ index.html
@@ -19,132 +19,7 @@
 			<h1>Quiz App</h1>
 		</header>
 		<main>
-			<section class="main__section" data-js="questionSection">
-				<section class="question-viewport">
-					<section class="questioncard">
-						<button class="button--bookmark bookmark" data-js="bookmarkIcon">
-							<img src="./assets/bookmark.svg" />
-						</button>
-						<h2 class="questioncard__header">
-							What is the correct way to write a comment in CSS?
-						</h2>
-						<button
-							class="button button--show-answer"
-							type="button"
-							data-js="answerButton">
-							Show Answer
-						</button>
-						<p class="answer answer--hidden" data-js="answerBlock">
-							The correct way to write a comment in CSS is /* comment */.
-						</p>
-						<ul class="taglist">
-							<li class="tag">#html</li>
-							<li class="tag">#css</li>
-							<li class="tag">#tag</li>
-						</ul>
-					</section>
-				</section>
-				<section class="question-viewport">
-					<section class="questioncard">
-						<button class="button--bookmark bookmark" data-js="bookmarkIcon">
-							<img src="./assets/bookmark.svg" />
-						</button>
-						<h2 class="questioncard__header">
-							What is the difference between an id selector and a class selector
-							in CSS?
-						</h2>
-						<button
-							class="button button--show-answer"
-							type="button"
-							data-js="answerButton">
-							Show Answer
-						</button>
-						<p class="answer answer--hidden" data-js="answerBlock">
-							An id selector is unique, while a class selector can be used
-							multiple times.
-						</p>
-						<ul class="taglist">
-							<li class="tag">#html</li>
-							<li class="tag">#css</li>
-							<li class="tag">#anotherTag</li>
-						</ul>
-					</section>
-				</section>
-				<section class="question-viewport">
-					<section class="questioncard">
-						<button class="button--bookmark bookmark" data-js="bookmarkIcon">
-							<img src="./assets/bookmark.svg" />
-						</button>
-						<h2 class="questioncard__header">
-							How can you include an external CSS file in an HTML page?
-						</h2>
-						<button
-							class="button button--show-answer"
-							type="button"
-							data-js="answerButton">
-							Show Answer
-						</button>
-						<p class="answer answer--hidden" data-js="answerBlock">
-							You can include an external CSS file in an HTML page by using the
-							&lt;link&gt; tag in the &lt;head&gt; section.
-						</p>
-						<ul class="taglist">
-							<li class="tag">#html</li>
-							<li class="tag">#css</li>
-						</ul>
-					</section>
-				</section>
-				<section class="question-viewport">
-					<section class="questioncard">
-						<button class="button--bookmark bookmark" data-js="bookmarkIcon">
-							<img src="./assets/bookmark.svg" />
-						</button>
-						<h2 class="questioncard__header">
-							Which attribute do you need to use to insert an image in HTML?
-						</h2>
-						<button
-							class="button button--show-answer"
-							type="button"
-							data-js="answerButton">
-							Show Answer
-						</button>
-						<p class="answer answer--hidden" data-js="answerBlock">
-							You need to use the src attribute to insert an image in HTML.
-						</p>
-						<ul class="taglist">
-							<li class="tag">#html</li>
-							<li class="tag">#css</li>
-							<li class="tag">#somanytags</li>
-						</ul>
-					</section>
-				</section>
-				<section class="question-viewport">
-					<section class="questioncard">
-						<button class="button--bookmark bookmark" data-js="bookmarkIcon">
-							<img src="./assets/bookmark.svg" />
-						</button>
-						<h2 class="questioncard__header">
-							How can you create a hyperlink in HTML that refers to another
-							webpage?
-						</h2>
-						<button
-							class="button button--show-answer"
-							type="button"
-							data-js="answerButton">
-							Show Answer
-						</button>
-						<p class="answer answer--hidden" data-js="answerBlock">
-							You can create a hyperlink in HTML that refers to another webpage
-							by using the &lt;a&gt; tag and the href attribute.
-						</p>
-						<ul class="taglist">
-							<li class="tag">#html</li>
-							<li class="tag">#css</li>
-							<li class="tag">#flexbox</li>
-						</ul>
-					</section>
-				</section>
-			</section>
+			<section class="main__section" data-js="questionSection"></section>
 			<section
 				class="main__section--hidden main__Section"
 				data-js="bookmarkSection"
